+++ src/features/activity/main/ui/activity-card/ActivityCard.tsx
