+++ src/features/activity/main/ui/category-filter/CategoryFilter.tsx
